+++ api/src/main/java/cn/qihangerp/api/controller/GoodsSupplierController.java
@@ -1,11 +1,9 @@
 package cn.qihangerp.api.controller;
 
 import cn.qihangerp.api.common.*;
-import cn.qihangerp.api.domain.ErpGoodsCategory;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import cn.qihangerp.api.domain.ErpGoods;
 import cn.qihangerp.api.domain.ErpSupplier;
 import cn.qihangerp.api.service.ErpSupplierService;
 
@@ -16,13 +14,15 @@ import java.util.List;
 @AllArgsConstructor
 @RestController
 @RequestMapping("/supplier")
-public class SupplierController extends BaseController{
+public class GoodsSupplierController extends BaseController{
     private final ErpSupplierService supplierService;
 
     @GetMapping("/list")
     public TableDataInfo list(ErpSupplier bo, PageQuery pageQuery)
     {
-        bo.setTenantId(getUserId());
+        if(getUserId()!=1) {
+            bo.setTenantId(getUserId());
+        }
         PageResult<ErpSupplier> pageResult = supplierService.queryPageList(bo, pageQuery);
         return getDataTable(pageResult);
     }
@@ -47,6 +47,9 @@ public class SupplierController extends BaseController{
     @PostMapping
     public AjaxResult add(@RequestBody ErpSupplier bo)
     {
+        if(getUserId()==1) {
+            return AjaxResult.error("超级管理员账号不允许操作");
+        }
         bo.setTenantId(getUserId());
         bo.setCreateBy("手动添加");
         bo.setCreateTime(new Date());
@@ -56,6 +59,9 @@ public class SupplierController extends BaseController{
     @PutMapping
     public AjaxResult edit(@RequestBody ErpSupplier bo)
     {
+        if(getUserId()==1) {
+            return AjaxResult.error("超级管理员账号不允许操作");
+        }
         bo.setTenantId(null);
         bo.setCreateTime(null);
         bo.setCreateBy(null);
@@ -67,6 +73,9 @@ public class SupplierController extends BaseController{
     @DeleteMapping("/{ids}")
     public AjaxResult remove(@PathVariable Long[] ids)
     {
+        if(getUserId()==1) {
+            return AjaxResult.error("超级管理员账号不允许操作");
+        }
         return toAjax(supplierService.removeBatchByIds(Arrays.stream(ids).toList()));
     }
